Change game id to long

+++ feature/game/model/src/commonMain/kotlin/com/flaringapp/ligretto/feature/game/model/GameId.kt
@@ -3,7 +3,7 @@ package com.flaringapp.ligretto.feature.game.model
 import kotlin.jvm.JvmInline
 
 @JvmInline
-value class GameId(val value: Int) {
+value class GameId(val value: Long) {
 
     companion object {
         fun zero() = GameId(0)
